+++ sampleAppKotlin/src/main/java/com/virtusize/sampleappkotlin/MainActivity.kt
@@ -1,0 +1,42 @@
+package com.virtusize.sampleappkotlin
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import com.virtusize.android.data.local.VirtusizeError
+import com.virtusize.android.data.local.VirtusizeEvent
+import com.virtusize.android.data.local.VirtusizeMessageHandler
+import com.virtusize.android.data.local.VirtusizeOrder
+import com.virtusize.android.data.local.VirtusizeOrderItem
+import com.virtusize.android.data.local.VirtusizeProduct
+import com.virtusize.android.data.local.VirtusizeViewStyle
+import com.virtusize.android.util.dpInPx
+import com.virtusize.android.util.spToPx
+import com.virtusize.sampleappkotlin.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var navController: NavController
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        // Get the navigation host fragment from this Activity
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        // Instantiate the navController using the NavHostFragment
+        navController = navHostFragment.navController
+        // Make sure actions in the ActionBar get propagated to the NavController
+        setupActionBarWithNavController(navController)
+    }
+
+    /**
+     * Enables back button support. Simply navigates one element up on the stack.
+     */
+    override fun onSupportNavigateUp(): Boolean {
+        return navController.navigateUp() || super.onSupportNavigateUp()
+    }
+}