+++ virtusize/src/main/java/com/virtusize/android/data/parsers/VirtusizeEventJsonParser.kt
@@ -1,0 +1,18 @@
+package com.virtusize.android.data.parsers
+
+import com.virtusize.android.data.local.VirtusizeEvent
+import org.json.JSONObject
+
+/**
+ * This class parses a JSONObject to the [VirtusizeEvent] object
+ */
+internal class VirtusizeEventJsonParser : VirtusizeJsonParser<VirtusizeEvent> {
+    override fun parse(json: JSONObject): VirtusizeEvent? {
+        val name = JsonUtils.optString(json, FIELD_NAME)
+        return VirtusizeEvent(name, json)
+    }
+
+    companion object {
+        private const val FIELD_NAME = "eventName"
+    }
+}