+++ virtusize/src/main/java/com/virtusize/android/data/remote/ProductCheck.kt
@@ -1,0 +1,17 @@
+package com.virtusize.android.data.remote
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+/**
+ * This class represents the response for the API request ProductCheck
+ * @see VirtusizeEndpoint.ProductCheck
+ */
+@Parcelize
+data class ProductCheck(
+    val data: Data?,
+    val productId: String,
+    val name: String,
+    // the JSON response as a String (For the Flutter SDK)
+    val jsonString: String?
+) : Parcelable